Update v2.0.14
New Methods:
- get_orders_history
- get_task_log

Fixes:
- Fixed a UTF-8 issue in the internal HTTP service.
- Fixed the request schema in store_git_credentials()

+++ src/js7/model/public/client/filter/resume_order_filter.py
@@ -41,5 +41,5 @@ class ResumeOrderFilter(BaseModel):
     An object with key-value pairs. The value can be a string, number or boolean.
     """
 
-    cycle_end_time: Optional[float] = None
+    cycle_end_time: Optional[int] = None
     """A relative cycle end time in seconds."""
